% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeObjects.R
\name{edgeObjects}
\alias{edgeObjects}
\title{Identify Objects along Edge of Image}
\usage{
edgeObjects(x, border = 1)
}
\arguments{
\item{x}{An \code{Image} object in \code{Grayscale} color mode or an
array containing object masks. Object masks are sets of pixels with
the same unique integer value.}

\item{border}{Number of pixels from the edge to include.}
}
\value{
A vector of integers or list of integers identifying objects at the edge
}
\description{
Identify objects within \code{border} pixels of the edge of an
integer \code{Image} object.
}
\examples{
  x <- readImage(system.file("extdata", "by_folder/b4/file001.tif", package = "virustiter"))
  nm0 <- nucMask(x)
  sel <- edgeObjects(nm0, border = 16)
  nm1 <- rmObjects(nm0, sel)
  plot(colorLabels(combine(nm0, nm1)), all = TRUE, nx = 1)

}
