% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCut.R
\name{getCut}
\alias{getCut}
\title{Determine Cutoff Value between Positive and Negative}
\usage{
getCut(df, by = c("control", "file", "well", "row", "column"), param = "y",
  mult = 5, log = TRUE)
}
\arguments{
\item{df}{Annotated \code{data.frame} with fluorescent values to evaluate.}

\item{by}{Character string identifying the group in which to seek cutoff
values \code{("control", "file", "well", "row" or "column")}.}

\item{param}{Variable name as character string in \code{df} to evaluate, 
typically \code{"y"}.}

\item{mult}{Muliplier constant passed to \code{findBgnd()}.}

\item{log}{\code{logical} flag passed to \code{findBgnd()} to use
log-transformed values.}
}
\value{
A named numeric vector of cutoff values.
}
\description{
Determine the cutoff value between "positive" and "negative" values in
argument \code{param}. The background values are specified by \code{by} as
control values or by row, column or file.
}
\details{
The cutoff value between positive and negative values in 
\code{param} will be determined according to \code{by}. If this
value is "control", all values identified as 
\code{type == "control"} or with \code{x == 0} will be considered to be 
background. Otherwise, a cutoff will be determined by Otsu's method
for the groups identified identified in \code{by} ("file", "well",
"column" or "row"). The cutoff value will be determined by the logic
in \code{findBgnd()}.

The annotated data frame must have the variable identified in 
\code{param} and a variable named \code{x} as well as 
\code{file, well, row} or 
\code{column} as appropriate.
}
