% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCutoff.R
\name{getCutoff}
\alias{getCutoff}
\title{Get cutoff values for nuclear mask area}
\usage{
getCutoff(mask, lo, hi, breaks = 256, quant = 0.995, plot = FALSE)
}
\arguments{
\item{mask}{An \code{Image} object, array, or a \code{list} of
these objects containing a binary mask defining regions of the image
to be segmented. Typically, this would obtained from
\code{\link[virustiter]{nucMask}}.}

\item{lo}{A optional numeric value indicating the approximate minimum value.
This value will be used to search for the true saddle point. If negative,
the search will start at the lower 10 percentile of the area.}

\item{hi}{A optional numeric value for the expected upper limit of the
cutoff. If missing, the\code{quantile} will be used to determine the
upper limit.}

\item{breaks}{Number of bins used for the area histograms (default of 256).}

\item{quant}{If \code{hi} is not provided, the area given by this
quantile will be used as the upper limit.}

\item{plot}{If \code{TRUE}, plot histogram with selected cutoff points}
}
\value{
Vector of two values recommended as nuclear mask cutoff parameters.
}
\description{
Find appropriate range limits for the area of nuclei defined in a
binary nuclear mask. The limits will be defined by the "saddle"
point on the low end and by the upper quantile on the high end.
}
\details{
Get the low and high cutoffs to give to \code{\link[virustiter]{trimMask}}.
}
