% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImages.R
\name{getImages}
\alias{getImages}
\title{Get Paired Microscopic Images}
\usage{
getImages(source, type = "tiff", which.images = c(1, 2, 2),
  pattern = NULL, verbose = FALSE)
}
\arguments{
\item{source}{A character vector identifying a directory or directories
with multilayer tiff files \emph{or} subdirectories identified
by well with separate, paired images \emph{or} a character vector
of image files \emph{or} a \code{.zip} file with the above.}

\item{type}{A character string identifying the type of image files to parse
("tif", "tiff", "jpeg", "jpg" or "png".)}

\item{which.images}{An integer of length 2 or 3. The first two numbers indicate
the relative position of the DNA image and the target in each field. The optional
third number specifies the total number of images for each field. A value of
\code{c(1, 2)} indicates DNA first and target second. A value of \code{c(2, 1)}
indicates that the order is target first and DNA image second. A value of
\code{c(1, 2, 3)} indicates a DNA image, a target image, and a third (ignored)
image such as a phase contrast image or second fluorescent color in each set.}

\item{pattern}{Optional character string to serve as a \code{grep} pattern
for \code{list.files()} to select image files.}

\item{verbose}{If \code{TRUE}, print diagnostic messages as files are read.}
}
\value{
A list of two named lists containing the nuclear and target images.
Each well or file will be represented as an element in the lists \code{nuc}
and \code{tgt}. Diagnostic messages are provided if \code{verbose}
is \code{TRUE}.
}
\description{
Return a list of paired DNA and fluorescent images
appropriate for \code{parseImages()}.
}
\details{
Images specified in \code{source} will be checked for the proper
organization of image files. The order required by \code{parseImages()}
is the DNA image first and fluorescent viral target second. This
order can be changed with the \code{which.images} argument.

Images associated with each multiplicity of infection can be individual
files in a single directory where each directory named as the well such as
\code{A1}, \code{A2}, etc. and the files within are identified as
\code{A1/file001.tif}, \code{A1/file002.tif}, etc. The well identifier
can be in upper or lower case and can contain leading zeros such as
\code{c0003/file12.tif}.

Alternatively, each group of images associated with a given moi can be
a multi-layered tiff file where the sequence of images in the file is
specified by the argument \code{which.images}.
}
\examples{
# Example with data organized by folder or well
  path.by.folder <- system.file("extdata", "by_folder", package = "virustiter")
  getImages(path.by.folder)

}
