% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTiter.R
\name{getTiter}
\alias{getTiter}
\title{Get Titer from Fitted Model}
\usage{
getTiter(fm, level = 0.95, digits = 2, p = 0.63212)
}
\arguments{
\item{fm}{fitted model \emph{or} list of fitted models generated by
\code{\link{getFit}}.}

\item{level}{this specifies the confidence level (default of 0.95).
Use \code{NULL} to return only the titer (or EC63 value).}

\item{digits}{\code{Integer} number of significant digits to report
(default = 2)}

\item{p}{probability (infected fraction) corresponding to an MOI of 1
(default = 1 - exp(-1) = 0.6321)}
}
\value{
If \code{level} was provided, the return value includes the estimate and
confidence intervals. Note that the confidence interval reflects the
precision of the binomial fit is not necessarily \emph{biologically}
meaningful.
}
\description{
Extract the viral titer \emph{aka} the volume required for one
infectious unit with confidence intervals from a fit object or
a list of fit objects generated by \code{\link{getFit}}.
}
