% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZero.R
\name{getZero}
\alias{getZero}
\title{Find Background (Zero Value) Pixel}
\usage{
getZero(img, frac = 0.5)
}
\arguments{
\item{img}{An array object or list of array objects, where each
is typically a grayscale \code{Image} object.}

\item{frac}{Number between 0 and 1 specifying the fraction of
values to consider.}
}
\value{
A numeric vector or list of numeric vectors with the most
common non-zero pixel for each frame.
}
\description{
Find the most common non-zero pixel in the lower fraction
of each frame of an image.
}
\details{
This function is meant to be applied to images that have more
background pixels than foreground pixels. The most common
value among the lower half (as specified by \code{frac})
of intensity values is sought as the true 'zero' value.

Intensity values greater than the minimum and less than
\code{frac * maximum} are processed by the \code{\link{density}}
function to determine the most common pixel. The minimum pixel
value is excluded to avoid skewing the density estimate if an
undersaturated image is used.
}
\seealso{
\code{\link{bnormalize}}, \code{\link{setZero}},
  \code{\link{getBgnd}}, \code{\link{findBgnd}}
}
