% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listImages.R
\name{listImages}
\alias{listImages}
\title{List Image Files}
\usage{
listImages(path = ".", type = c("tiff", "jpeg", "jpg", "png"),
  pattern = NULL, full.names = TRUE, recursive = TRUE,
  ignore.case = TRUE)
}
\arguments{
\item{path}{A character vector of path names.}

\item{type}{The type of image as defined by the file extension (tiff, jpeg, jpg or png).}

\item{pattern}{Argument passed to list.files (grep pattern), default is NULL.}

\item{full.names}{Argument passed to list.files, default is TRUE.}

\item{recursive}{Argument passed to list.files, default is TRUE.}

\item{ignore.case}{Argument passed to list.files, default is TRUE.}
}
\value{
Character vector of selected image filenames.
}
\description{
List image files in the given path.
}
\examples{
  path.by.folder <- system.file("extdata", "by_folder", package = "virustiter")
  listImages(path.by.folder, pattern = "002")

}
