% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.plate.R
\name{make.plate}
\alias{make.plate}
\title{Create data.frame with Tag, Prefix, Well, Column and Row}
\usage{
make.plate(n, prefix = NULL, digits = NULL, upper = TRUE, byrow = TRUE)
}
\arguments{
\item{n}{the number of wells (6, 12, 24, 48, 96, or 384)}

\item{prefix}{optional \code{character} vector of prefix values}

\item{digits}{optional \code{integer} value to zero-pad the column labels}

\item{upper}{\code{logical} flag to use upper case for row labels}

\item{byrow}{\code{logical} value specifying organization}
}
\value{
A data.frame of length three or five with the following items. Note that 
values for \code{tag} and \code{prefix} will be returned only if
a non-NULL value for prefix is provided. 
\itemize{
 \item tag, harmonized character representation of the entire well name
   including any prefix, if a prefix was provided
 \item prefix, prefix as a character string, if provided
 \item well, harmonized well name as a factor
 \item row, row as a factor 
 \item column, column number as a factor
}
}
\description{
Create a data.frame representing a multi-well plate with an optional 
prefix.
}
\details{
Generate a data.frame suitable for typical phenodata. The \code{tag} 
is formed by replicating the prefix and pasting it in
front of each well value.
}
\examples{
  make.plate(6)
  make.plate(6, upper = FALSE)
  make.plate(12, 1:3)
  make.plate(24, 1:6)$tag # typical use to generate "OTE" labels

}
