% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOneFit.R
\name{plotOneFit}
\alias{plotOneFit}
\title{Plot One Titer Fit with more Options}
\usage{
plotOneFit(fm, main = NULL, xlab = NULL, ylab = NULL, pch.col = "black",
  line.col = "red", ref.col = "blue", ...)
}
\arguments{
\item{fm}{Fitted model or list of fitted models from \code{getFit()}.}

\item{main, xlab, ylab, pch.col}{Parameters passed to \code{plot}.}

\item{line.col}{Color for best-fit line (appropriate for \code{par("col")}.
param ref.col Color for reference lines identifying key Poisson points.}

\item{...}{Additional arguments passed to \code{plot()}.}

\item{by}{Character vector indicating that fits are organized as a
single fit ("none") or by "column" or "row".}
}
\value{
No value is returned. This function is called for the side-effect of 
producing a plot with base graphics.
}
\description{
Prepare a single diagnostic plot with more labeling options than \code{plotFit()}.
}
\details{
Base graphics are used to prepare the plot where the first argument is
a single fitted model from \code{getFit()}. The function calls
\code{getTiter()} to obtain the fit and confidence intervals.
}
