% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWell.R
\name{plotWell}
\alias{plotWell}
\title{Plot Results by Well with Lattice Graphics}
\usage{
plotWell(df, myWell = levels(df$well), byFrame = TRUE, cex = 1,
  invert.y = TRUE, ...)
}
\arguments{
\item{df}{Annotated \code{data.frame} with imaging results.}

\item{myWell}{Character vector of wells to be plotted.}

\item{byFrame}{A \code{logical} value to separate frames.}

\item{cex}{Numeric value as factor by which to expand the plot symbol.}

\item{invert.y}{A \code{logical} value to invert y coordinates.}

\item{...}{Additional arguments handed to \code{xyplot}.}
}
\value{
The plot is returned as an invisible \code{lattice} object.
}
\description{
Plot a representation of the selected wells showing positive and
negative cells with \code{\link[lattice]{xyplot}}.
}
