% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIJResults.R
\name{readIJResults}
\alias{readIJResults}
\title{Read Results from ImageJ Multiwell "Fluorescent Cell Count"}
\usage{
readIJResults(f, SEP = NULL)
}
\arguments{
\item{f}{Path to ImageJ results file, single character vector.}

\item{SEP}{Character used to separate directory and filename in ImageJ
results file.}
}
\description{
Legacy code to read results from ImageJ code that performs the work now done in
\code{parseImages}.
}
\details{
Not recommended to use this any more. It has not been well maintained
since early 2016. At last incarnation, this program expects variables
from ImageJ as "Area", "Mean gray value", "Center of Mass", "Display
label", and "Inverted Y coordinates". "Label" must be of the form
"xnn/filename" where xnn is the well such as "A01" or "b2". The ImageJ
results file should have the multiplicity as an x or moi value and a
unit indicator (unit) coded as 1, 2, 3 or 4 for "VP per cell", "IU per
cell", "ul per cell", "ml per cell".
}
