% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Score Cells as Positive/Negative}
\usage{
score(df, cut = NULL, param = "mfi", ...)
}
\arguments{
\item{df}{A data.frame generated by \code{parseImages()} or
\code{readIJResults()} and appropriate grouping variable.}

\item{cut}{A numeric vector of length one or a named numeric vector for
each level of the grouping variable \code{well, row, column, or file}. If
missing, the function \code{getCut} will be called with default parameters
to determine the cutoff.}

\item{param}{A character string identifying the variable to be scored.}

\item{...}{Additional arguments for \code{getCut()} such as \code{by}.}
}
\description{
Assign \code{logical} variable \code{positive} as TRUE or FALSE in data.frame
of fluorescent intensity values based on cutoff values in \code{cut}.
}
\details{
A variable named \code{positive} will be added to the data.frame if 
necessary and assigned \code{TRUE} or \code{FALSE} if the value 
identified as "param" is greater than the appropriate value in
\code{cut}. If \code{cut == NULL}, \code{getCut()} will be called
with any additional arguments in \code{...} to determine the cutoff.
}
\examples{
# Small set of data by stack
  f <- system.file("extdata", "by_stack/file005.tif", package = "virustiter")
  file.copy(f, tempdir(), overwrite = TRUE)
  v <- parseImages(tempdir(), pattern = "file005")

# Score with grouping value of "file"
  v <- score(v, by = "file")
  xtabs(~ positive, v)
  plot(log(mfi) ~ area, v, col = c(1,2)[positive + 1])

}
