% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setZero.R
\name{setZero}
\alias{setZero}
\title{Normalize Image to Given Zero Value}
\usage{
setZero(img, zero, bit.depth = 16, nonzero = TRUE)
}
\arguments{
\item{img}{An array object or list of array objects, where each
is typically a grayscale \code{Image} object.}

\item{zero}{Number between 0 and 1 for the new zero value.}

\item{bit.depth}{Integer between 1 and 64 specifying the
bit-depth of the image, typically 8 or 16.}

\item{nonzero}{Logical value, if \code{TRUE}, the minimum value will
adjusted to \code{2^-bit.depth}, otherwise the minimum value
will be set to 0.}
}
\value{
An array or list of the same size that has been linearly scaled.
}
\description{
Linearly scale an image to re-zero at the given zero value.
}
\details{
Each frame of the argument will be linearly scaled to have a
minimum at \code{zero} by subtraction. Negative values will
be clipped to 0 if \code{nonzero} is \code{FALSE} otherwise the
minimum value will be set to \code{2^-bit.depth}. 

This function will probably replace \code{\link{bnormalize}}
soon. This function is typically applied after using
\code{\link{getZero}} to determine the likely zero value.
}
\seealso{
\code{\link{bnormalize}}, \code{\link{getZero}}
}
