% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showImages.R
\name{showImages}
\alias{showImages}
\title{Show Paired Microscopic Images}
\usage{
showImages(path, type = "tiff", which.images = c(1, 2, 2), pattern = NULL,
  method = c("raster", "browser"))
}
\arguments{
\item{path}{A character vector of directory with either multilayer tiff
image files \emph{or} subdirectories identified by well with separate,
paired images per well.}

\item{type}{A character string identifying the type of image files to parse
("tif", "tiff", "jpeg", "jpg" or "png".)}

\item{which.images}{An integer of length 2 or 3. The first two numbers indicate
the relative position of the DNA image and the targetin each field. The optional
third number specifies the total number of images for each field. A value of
c(1, 2) indicates DNA first and target second. A value of c(2, 1) indicates that
the order is target first and DNA image second. A value of c(1, 2, 3)
indicates a DNA image, a target image, and a third (ignored) image such as a
phase contrast image or second fluorescent color in each set.}

\item{pattern}{Optional grep pattern as character string used by \code{list.files()}
to select image files.}

\item{method}{Character string specifying the method of displaying images.
Default of \code{"browser"} using web browser, \cdoe{"raster"} uses R
raster graphics.}
}
\value{
This function is called to display subsets of normalized image
files. Use the \code{grep} string in \code{pattern} to limit the selected
images. The last group of images displayed is invisibly returned.
}
\description{
Show paired DNA and a fluorescent images
from paired images using the same selection as \code{parseImages()}.
}
\details{
Images specified in \code{path} will be read, normalized and displayed
with the same logic in \code{parseImages()}. This displays the images by
either a browser or by raster and performs basic checks to ensure that
the proper number of files are present. This has been implemented
with \code{\link{EBImage}} and is part of a suite of tools to determine
viral titers from fluorescent micrograph pairs. Typically, the first of each
pair is a DNA image and the second a fluorescent image of the viral target.
target signal. Although the nuclear (typically DAPI) image file is expected
to precede the corresponding viral antigen image file, this order can be
changed with the \code{which.images} argument.

Pairs of images associated with each moi can be individual files in a
single directory where each directory is named for the well such as
\code{A1}, \code{A2}, etc. and the files within are identified as
\code{A1/file001.tif}, \code{A1/file002.tif}, etc. The well identifier
can be in upper or lower case and can contain leading zeros such as
\code{c0003/file12.tif}.

Alternatively, each group of images associated with a given moi can be
a multi-layered tiff file where the sequence of images in the file is
specified by the argument \code{which.images}.
}
\examples{
# Example with data organized by folder or well
  path.by.folder <- system.file("extdata", "by_folder", package = "virustiter")
  showImages(path.by.folder)

}
