% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally.R
\name{tally}
\alias{tally}
\title{Tally Positive Fluorescent Values}
\usage{
tally(df, pd = NULL, param = "mfi", moi = c("x", "moi"))
}
\arguments{
\item{df}{Annotated \code{data.frame} with fluorescent values to evaluate.}

\item{pd}{Optional phenotype \code{data.frame} to add to results.}

\item{param}{Variable name as character string in \code{df} to evaluate, 
typically \code{"mfi"}.}

\item{moi}{Character string identifying the independent value, typically
the multiplicity of infection or \code{"moi"}.}
}
\value{
A \code{data.frame} of tallied values with (optional) additional information
provided by \code{pd}.
}
\description{
Tally positive and negative values from from data.frame generated
by \code{parseImages()} by either "well" or "file" representing each
value of multiplicity.
}
\details{
The variable named \code{positive} will be tallied for each level of
"file" or "well" in the argument \code{df}. The dependent variable is
expected to by "moi" or "x".
}
\examples{
# Sample data by folder
  f <- system.file("extdata", "by_folder", package = "virustiter")
  pd <- read.csv(system.file("extdata", "by_folder/phenoData.csv", package = "virustiter"))
  v <- parseImages(f)
  v <- mergePdata(pd, v)
  v <- score(v)
  tally(v)

}
