% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Linear Spatial Translation}
\usage{
translate(x, adj, filter = c("bilinear", "none"), ...)
}
\arguments{
\item{x}{An \code{Image} object or array or list of the same.}

\item{adj}{A vector of 2 numbers denoting the translation vector
in pixels or a list of the same.}

\item{filter}{A character string indicating the interpolating
sampling filter. Valid values are \code{"none"} or the default, 
\code{"bilinear"}.}

\item{...}{Arguments to be passed to \code{affine} in the
\code{EBImage} package, such as \code{filter}, \code{output.dim},
\code{bg.col} or \code{antialias}.}
}
\description{
Replacement for \code{EBImage} function that accepts lists
}
\details{
This function is a replacement for the EBImage code of the same name.
}
