% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{usage}
\alias{usage}
\title{Show Usage of \code{virustiter} Package}
\usage{
usage()
}
\value{
No value returned, a window is opened with helpful information.
}
\description{
Display succinct summary of functions and work flow in \code{virustiter}.
}
