% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virustiter-package.R
\docType{package}
\name{virustiter-package}
\alias{virustiter}
\alias{virustiter-package}
\title{virustiter: Process fluorescent images for virus titer}
\description{
A suite of code that helps determining virus titers from fluorescent
micrographs, typically from paired fluorescent images of the cell nucleus
and a viral antigen.
}
\author{
\strong{Maintainer}: David Ornelles \email{david.ornelles@wfusm.edu}

}
\keyword{internal}
