% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOneFit.R
\name{addOneFit}
\alias{addOneFit}
\title{Add Single Fitted Model to Plot}
\usage{
addOneFit(fm, line.col = 4, ref.col = 4, pch.col = line.col, ...)
}
\arguments{
\item{fm}{Fitted model or list of fitted models from \code{\link{getFit}}.}

\item{line.col}{Color for best-fit line appropriate for \code{par("col")}.}

\item{pch.col}{Pch color passed to \code{plot}.}

\item{...}{Additional arguments passed to \code{\link{plot}}.}
}
\value{
No value is returned. This function is called for the side-effect of 
adding to an existing plot with \code{\link{points}} and \code{\link{lines}}.
}
\description{
Add a single fitted model plot to an existing plot made with base graphics
}
\details{
Prepare a single diagnostic plot with more labeling options than \code{plotFit()}.



Base graphics are used to add a plot to an existing plot produced by 
\code{\link{plotFit}} or \code{\link{plotOneFit}}. The first argument is
a single fitted model from \code{\link{getFit}}. The function calls
\code{\link{getEC63}} to obtain the fit and confidence intervals.
}
