% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOneFit.R
\name{addOneFit}
\alias{addOneFit}
\title{Add Single Fitted Model to Plot}
\usage{
addOneFit(
  fm,
  pch = 1,
  col = 2,
  col.pch = col,
  lty.fit = 1,
  col.fit = col,
  lty.ref = 2,
  col.ref = "gray",
  ...
)
}
\arguments{
\item{fm}{single fitted model from \code{\link{getFit}}.}

\item{pch, col, col.pch}{plot character and color for data points and the 
default color for the fitted line. Set \code{pch} to \code{NA}
to exclude plotting characters.}

\item{lty.fit, col.fit}{line type and color for the GLM best-fit line.  Set
\code{lty.fit} to \code{NA} to exclude the best-fit line.}

\item{lty.ref, col.ref}{line type and color for the value on the x-axis
that intersect the 63% value on the best-fit line. Set \code{lty.ref} to
\code{NA} to exclude the reference line.}

\item{...}{Additional arguments passed to \code{\link{plot}}.}
}
\value{
No value is returned. This function is called for the side-effect of 
adding to an existing plot with \code{\link{points}} and \code{\link{lines}}.
}
\description{
Add a single fitted model plot to an existing plot made with base graphics
}
\details{
Base graphics are used to add a plot to an existing plot produced by 
\code{\link{plotFit}}. The first argument is
a single fitted model from \code{\link{getFit}}. The function calls
\code{\link{getTiter}} to obtain the fit and confidence intervals.
}
