% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnormalize.R
\name{bnormalize}
\alias{bnormalize}
\title{Normalize Image Background}
\usage{
bnormalize(img, inputRange, quant = 0.025, nonzero = FALSE, ...)
}
\arguments{
\item{img}{Grayscale \code{Image} object.}

\item{inputRange}{A numeric vector of 2 values specifying the expected
range of intensity values.}

\item{quant}{Quantile to serve as the common baseline, default of 0.025.}

\item{nonzero}{Should zero values be ignored in quantile calculation?}

\item{...}{Additional arguments are accepted but ignored in order
to remain compatible with previous versions of this function.}
}
\value{
\code{Image} object linearly scaled to have a common background.
}
\description{
Adjust values of an image to a common background.
}
\details{
Each frame of an \code{Image} will be linearly scaled to have an
identical value at the specified quantile. If \code{inputRange} is missing,
it will be taken as the actual range of values in \code{img}.
Each frame of the image will be adjusted by subtracting the median value at
the specified quantile and adding \code{quant * inputRange[2]}. Negative 
values will be clipped to 0. \emph{Note that if \code{inputRange}
is not specified, the different image objects will not necessarily share
the same background value.}
}
\examples{
  cell <- readImage(system.file("extdata", "by_folder/b2/file002.tif", 
   package = "virustiter"))
  N <- prod(dim(cell)[1:2])
  cells <- Image(rep(imageData(cell), 4), dim = c(dim(cell), 4))
  cells[,,2] <- cell + 0.005
  cells[,,3] <- cell + 0.01
  cells[,,4] <- cell + 0.02
  cells <- normalize(cells, separate = FALSE)
  plot(cells, all = TRUE)
  plot(bnormalize(cells), all = TRUE)

}
