% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnormalize.R
\name{bnormalize}
\alias{bnormalize}
\title{Normalize Image Background}
\usage{
bnormalize(img, size = 2, sigma = 2, offset = 0.05)
}
\arguments{
\item{img}{\code{Image} grayscale image object}

\item{size}{Radius for \code{medianFilter}, integer. Default value of 2 
for typical images; use 0 to skip median filter.}

\item{sigma}{Standard deviation for \code{gblur}. Default value of 2 
for typical images, use 5 for finely detailed images; use 0 to skip 
Gaussian blur.}

\item{offset}{Value (default of 0.05) added to image after subtracting 
median value.}
}
\value{
\code{Image} object of the same size that has been smoothed, blurred 
and linearly adjusted.
}
\description{
Adjust images to have a common background value by simple linear shift.
}
\details{
This is meant to be applied to images with a greater number of pixels 
representing background values than foreground values. The median value 
of the image will be treated as the background and each image adjusted 
to have an identical median background value. The image(s) will first 
be smoothed by \code{medianFilter()} with radius \code{size} then blurred 
by \code{gblur} with \code{sigma}. Each processed image will be 
adjusted by subtracting the median value and adding \code{offset}.
}
\examples{
  cell <- readImage(system.file("extdata", "by_folder/b2/file002.tif", 
   package = "virustiter"))
  N <- prod(dim(cell)[1:2])
  cells <- Image(rep(imageData(cell), 4), dim = c(dim(cell), 4))
  cells[,,2] <- cell + 0.005
  cells[,,3] <- cell + 0.01
  cells[,,4] <- cell + 0.02
  cells <- normalize(cells, separate = FALSE)
  plot(cells, all = TRUE)
  plot(bnormalize(cells), all = TRUE)

}
