% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellMask.R
\name{cellMask}
\alias{cellMask}
\title{Voronoi Cell Segmentation from Nuclei}
\usage{
cellMask(seeds, mask = NULL, brush = NULL, lambda = 1e-04)
}
\arguments{
\item{seeds}{An \code{Image} object, array, or list of these objects
containing the seeds of identified regions. This would typically
be a segmented nuclear mask or list of masks such as those generated
by \code{\link{nucMask}}.}

\item{mask}{An optional \code{Image} object, array, or a \code{list} of
these objects containing a binary mask defining regions of the image
to be segmented. If this value is \code{NULL}, the nuclear mask
(\code{seeds}) will be expanded for Voronoi segmentation. If the
first argument \code{seeds} is a list, \code{mask} must be a
list of identical size with \code{Image} binary objects or arrays.}

\item{brush}{Size of the brush to expand the nuclear mask as an 
odd number of pixels. If this value is \code{NULL}, the mean value of 
the semi-major axis of the nuclei will be used.}

\item{lambda}{A numeric value used by \code{\link[EBImage]{propagate}}
determining the trade-off between the Euclidean distance in the image  
plane and the contribution of the gradient to define the boundary during
Voronoi segmentation. See \code{\link[EBImage]{propagate}} for details.}
}
\value{
An \code{Image} object produced by \code{\link[EBImage]{propagate}} 
containing the labeled objects (cells) or a \code{list} of such objects.
}
\description{
Define boundaries between adjacent cells (regions) from a
nuclear mask and optional cellular mask with the
\code{\link[EBImage]{propagate}} function.
}
\details{
A mask to define \emph{approximate} cellular boundaries will be created
from a nuclear mask in \code{seeds} and an optional cytoplasmic
mask \code{mask}. If the second argument (\code{mask}) is \code{NULL},
the nuclear mask will be dilated with a disc-shaped brush of size equal
to \code{brush} or, if \code{brush} is \code{NULL}, the average semi-major
axis of all nuclei. If \code{mask} is not \code{NULL}, \code{mask} must be
a binary (or integer) mask defining the limits for the Voronoi segmentation
based on the seeds provided in \code{seeds}. Such a binary mask can be
created by thresholding a non-specific widespread cytoplasmic signal such
antibody labeling for actin or a diffuse membrane stain.

To create a \emph{smaller} nuclear mask, use \code{\link{trimMask}} on a 
nuclear mask with a negative brush value.

To create a cytoplasmic mask that excludes the nucleus, generate a 
cell mask and combine this with the nuclear mask \code{nmask} as shown below. 

\preformatted{
# When both are single objects:
  cytoplasm <- cellMask(nmask) * (nmask == 0)

# When both are list objects:
  cytoplasm <- Map(function(a, b) a * (b == 0), cellMask(nmask), nmask)
}
}
\examples{
  x <- readImage(system.file("extdata", "by_folder/a4/file001.tif", package = "virustiter"))
  y <- readImage(system.file("extdata", "by_folder/a4/file002.tif", package = "virustiter"))
  nm <- nucMask(x)
  cm <- cellMask(nm)
  img <- rgbImage(red = normalize(y) * 0.2, green = normalize(y) * 0.8)
  img <- paintObjects(nm, img, col = "cyan")
  img <- paintObjects(cm, img, col = "red")
  plot(img)

}
