% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkImages.R
\name{checkImages}
\alias{checkImages}
\title{Check Paired Microscopic Images}
\usage{
checkImages(source, type = "tiff", which.images = c(1, 2, 2),
  pattern = NULL, method = c("none", "raster", "browser"),
  ask = TRUE, separate = FALSE, nx = 2)
}
\arguments{
\item{source}{A character vector identifying a directory or directories
with multilayer tiff files \emph{or} subdirectories identified
by well with separate, paired images per well \emph{or} a character vector
of image files \emph{or} a \code{.zip} file with the above.}

\item{type}{A character string identifying the type of image files to parse
("tif", "tiff", "jpeg", "jpg" or "png".)}

\item{which.images}{An integer of length 2 or 3. The first two numbers indicate
the relative position of the DNA image and the target image in each field. The optional
third number specifies the total number of images for each field. A value of
\code{c(1, 2)} indicates DNA first and target second. A value of \code{c(2, 1)}
indicates that the order is target first and DNA image second. A value of
\code{c(1, 2, 3)} indicates a DNA image, a target image, and a third (ignored)
image such as a phase contrast image or second fluorescent color in each set.}

\item{pattern}{Optional grep pattern as character string used by \code{\link{list.files}}
to select image files.}

\item{method}{Character string specifying the method of displaying images.
Default of \code{"none"} simply summarizes the images. A value of \code{"raster"}
uses R raster graphics and \code{"browser"} attempts to use a browser.
(Unfortunately this seems to be failing with EBImage version 4.22.0.)}

\item{ask}{Logical value to use \code{par(ask = TRUE)} if \code{method = "raster"}.}

\item{separate}{Logical value to normalize each frame separately if
\code{TRUE}. Aplies only if \code{display} is not \code{"none"}.}

\item{nx}{Integer value passed to \code{display{}} that specifies the number of
images in a column if \code{display} is not \code{"none"}.}
}
\value{
An \code{invisible} character vector of image files, diagnostic messages
are printed on the console with the option to display normalized image pairs.
}
\description{
Check validity of paired DNA and a fluorescent images
appropriate for \code{parseImages()}.
}
\details{
Images specified in \code{source} will be evaluated with the same logic in
\code{\link{getImages}} to determine if the proper number of files are
present and report on the number and form of the image files. 
The default value of the \code{which.images} argument treats the first of each
pair a DNA image and the second as a fluorescent 
image of the viral target.

Images associated with each multiplicity of infection can be individual
files in a single directory where each directory named as the well such as
\code{A1}, \code{A2}, etc. and the files within are identified as
\code{A1/file001.tif}, \code{A1/file002.tif}, etc. The well identifier
can be in upper or lower case and can contain leading zeros such as
\code{c0003/file12.tif}.

Alternatively, each group of images associated with a given moi can be
a multi-layered tiff file where the sequence of images in the file is
specified by the argument \code{which.images}.

If \code{'source'} is a zip file, files in the temporary directory 
(\code{\link{tempdir}}) will be deleted in order to receive the compressed
files.
}
\examples{
# Example with data organized by folder or well
  path.by.folder <- system.file("extdata", "by_folder", package = "virustiter")
  checkImages(path.by.folder)

}
