% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBgnd.R
\name{findBgnd}
\alias{findBgnd}
\title{Find Background Value by Otsu's Method}
\usage{
findBgnd(x, mult = 2.5, log = TRUE)
}
\arguments{
\item{x}{Fluorescent values to evaluate.}

\item{mult}{Numeric multiplier applied to the standard deviation,
with default value of 2.5.}

\item{log}{A \code{logical} flag to use log-transformed values.}
}
\value{
The value estimated as position of the background peak + mult * standard
deviation of the normal population.
}
\description{
Calculate a value between background and foreground pixels using
an approximation of Otsu's method for a mixture of values with a
normal density. This function assumes the background population
has a normal distribution and lies on the left of distribution.
}
\details{
Data are assumed to be drawn from two unimodal, continuous 
distributions where the background population follows a normal distribution. 
The maximum of this population is determined from a kernel density estimate
and the \emph{left} half of the distribution is fit to a Gaussian
distribution. The value returned is the position of the peak +
\code{mult} times the standard deviation of the distribution.
Note that fluorescent values are typically log-transformed before analysis.
The parameter \code{mult} must be empirically determined.

This code replaces a previous version that used the \code{half.range.mode()}
function in the \code{genefilter} package.
}
\examples{
  x <- c(rlnorm(200), rlnorm(120, 4))
  plot(density(log(x)))
  abline(v = log(findBgnd(x, mult = 0)), col = 4) # peak
  abline(v = log(findBgnd(x)), col = 2) # background limit

}
