% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBgnd.R
\name{findBgnd}
\alias{findBgnd}
\title{Find Cutoff by Otsu's Method}
\usage{
findBgnd(x, mult = 5, log = TRUE, left = TRUE)
}
\arguments{
\item{x}{Fluorescent values to evaluate.}

\item{mult}{Arbitrary numeric multiplier applied to standard deviation.}

\item{log}{\code{logical} flag to use log-transformed values.}

\item{left}{\code{logical} flag to indicate that the background population lies to the
left (\code{TRUE}) of the distribution.}
}
\value{
The cutoff estimate as half.mode + mult * standard deviation of the normal population.
}
\description{
Calculate a working cutoff value using Otsu's method in the package
\code{genefilter} for a mixture of values with a normal density. The
method assumes the background population lies to the left
of the normal mean of the population. This can be changed by the parameter
\code{left}.
}
\details{
Data are assumed to be drawn from two unimodal, continuous 
distributions where the background population of a normal distribution 
lies on the left. The mode of this population is estimated by an 
iterative, half-range method. The cutoff value returned is the mode +
\code{mult} * standard deviation. Note that fluorescent values are
typically log-transformed before analysis. The \code{mult} must be
empirically determined. The default value of 5 is fairly conservative.
}
\examples{
x <- c(rnorm(1000), rnorm(500, mean = 3))
findBgnd(x, mult = 0, log = FALSE)
plot(density(x))
abline(v = findBgnd(x, mult = 2, log = FALSE))

}
