% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findObjects.R
\name{findObjects}
\alias{findObjects}
\title{Find Objects}
\usage{
findObjects(expr, data, mask, invert = FALSE)
}
\arguments{
\item{expr}{An \code{expression} evaluated in \code{'data'} that returns
a logical vector identifying objects (rows) in \code{'data'}.}

\item{data}{A \code{data.frame} produced by \code{parseImages()} in which
\code{'expr'} will be evaluated.}

\item{mask}{An optional \code{Image} object in \code{Grayscale} mode or
an array containing object masks. Object masks are sets of pixels with
the same unique integer value.}

\item{invert}{A \code{logical} value to invert the selection determined
by \code{'expr'}.}
}
\value{
An integer \code{Image} mask or list of integer \code{Image} masks with the
objects selected by \code{'expr'} if \code{mask} was provided otherwise
a nested list of integers identifying the selected objects.
}
\description{
Identify objects in data.frame produced by \code{parseImages()} 
as specified by the expression \code{expr}.
}
\examples{
  x <- getImages(system.file("extdata", "by_folder/b4", package = "virustiter"))
  nm0 <- nucMask(x$nuc)
  df0 <- parseImages(x$nuc, x$tgt, nm0)
  nm1 <- findObjects(area > 190 & area < 220, df0, nm0)
  a <- Map(paintObjects, nm1, lapply(x$nuc, function(v) toRGB(normalize(v))))
  plot(a$b4, all = TRUE)

}
