% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAIC.R
\name{getAIC}
\alias{getAIC}
\title{Obtain AIC for 'Cut' Value}
\usage{
getAIC(df, cut, by = c("sequential", "column", "row"))
}
\arguments{
\item{df}{Annotated \code{data.frame} or list of \code{data.frame}s with
imaging results.}

\item{cut}{Numeric vector of cutoff values to test.}

\item{by}{A character string indicating the organization of the results as
being a single result or by column or by rows.}
}
\value{
The AIC value associated with the fit for each value of cut or list of AIC
values if a list of data frames was provided.
}
\description{
Apply a give \code{cut} value to create a fitted model (or models) and
return the Akaike's Criterion for the GLM fitted model. \code{getFit()}
is called for each value of \code{cut}.
}
