% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAIC.R
\name{getAIC}
\alias{getAIC}
\title{Obtain AIC for Given Background Value}
\usage{
getAIC(df, bgnd, by = c("sequential", "column", "row"))
}
\arguments{
\item{df}{Annotated \code{data.frame} or list of \code{data.frame}s with
imaging results.}

\item{bgnd}{Numeric vector of background values to test.}

\item{by}{A character string indicating the organization of the results as
being a single result or by column or by rows.}
}
\value{
The AIC value associated with the fit for each value of bgnd or list of AIC
values if a list of data frames was provided.
}
\description{
Apply a given \code{bgnd} value to create a fitted model (or models) and
return the Akaike's Criterion for the GLM fitted model. \code{\link{getFit}}
is called for each value of \code{bgnd}.
}
