% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBgnd.R
\name{getBgnd}
\alias{getBgnd}
\title{Determine Value Discriminating Positive from Negative Values}
\usage{
getBgnd(df, by = c("control", "file", "well", "row", "column"),
  param = "mfi", mult = 2.5, log = TRUE)
}
\arguments{
\item{df}{Annotated \code{data.frame} with fluorescent values to evaluate.}

\item{by}{Character string identifying the group in which to seek background
values, typically in \code{("control", "well", "file", "row" or "column")}.}

\item{param}{Variable name as character string in \code{df} to evaluate, 
typically \code{"y"}.}

\item{mult}{Muliplier constant passed to \code{findBgnd()}.}

\item{log}{\code{logical} flag passed to \code{findBgnd()} to use
log-transformed values.}
}
\value{
A named numeric vector of background values.
}
\description{
Determine the optimum background value between "positive" and "negative"
values in the argument \code{param}. The group used to determine the
background is specified by the argument \code{by}.
}
\details{
The value between positive and negative values in 
\code{param} will be determined according to \code{by}. If this
value is "control", all values identified as 
\code{type == "control"} or with \code{x == 0} will be considered to be 
background. Otherwise, a background will be determined by Otsu's method
for the groups identified identified by the character string in \code{by}.
Typically his would be \code{"file"} or \code{"well"} but can be any factor
variable in the data.frame \code{df}. The background value will be
determined by the logic in \code{findBgnd()}.

The annotated data frame must have the variable identified in 
\code{param} and a variable named \code{x} as well as an appropriate
grouping value identified in \code{by}.
}
\examples{
# Subset of data by_stack
  f <- system.file("extdata", "by_stack/file005.tif", package = "virustiter")
  img <- getImages(f)
  v <- parseImages(img)

# Get background value for grouping value of "file"
  getBgnd(v, by = "file")

}
