% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBgnd.R
\name{getBgnd}
\alias{getBgnd}
\title{Determine Value Discriminating Positive from Negative Values}
\usage{
getBgnd(df, by = c("control", "file", "well", "row", "column"),
  param = "mfi", mult = 2.5, log = TRUE)
}
\arguments{
\item{df}{Annotated \code{data.frame} with fluorescent values to evaluate.}

\item{by}{Character string identifying the group in which to seek background
values, typically in \code{("control", "file", "well", "row" or "column")}.}

\item{param}{Variable name as character string in \code{df} to evaluate, 
typically \code{"y"}.}

\item{mult}{Muliplier constant passed to \code{findBgnd()}.}

\item{log}{\code{logical} flag passed to \code{findBgnd()} to use
log-transformed values.}
}
\value{
A named numeric vector of background values.
}
\description{
Determine the maximum background value betwen "positive" and "negative"
values in argument \code{param}. The group used to determine the background
is specified by \code{by} as control values or by row, column or file.
}
\details{
The value between positive and negative values in 
\code{param} will be determined according to \code{by}. If this
value is "control", all values identified as 
\code{type == "control"} or with \code{x == 0} will be considered to be 
background. Otherwise, a background will be determined by Otsu's method
for the groups identified identified in \code{by} ("file", "well",
"column" or "row"). The background value will be determined by the logic
in \code{findBgnd()}.

The annotated data frame must have the variable identified in 
\code{param} and a variable named \code{x} as well as an appropriate
grouping value: either \code{file, well, row} or \code{column}
as required.
}
\examples{
# Subset of data by_stack
  f <- system.file("extdata", "by_stack/file005.tif", package = "virustiter")
  file.copy(f, tempdir(), overwrite = TRUE) # need files to reside in directory
  v <- parseImages(tempdir(), pattern = "file005")

# Get background value for grouping value of "file"
  getBgnd(v, by = "file")

}
