% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCutoff.R
\name{getCutoff}
\alias{getCutoff}
\title{Get Cutoff Values for Nuclear Mask Area}
\usage{
getCutoff(area, lo, hi, breaks = 256, quant = 0.995, plot = FALSE)
}
\arguments{
\item{area}{either an \code{Image} object, array, or a \code{list} of
these objects containing a binary mask defining regions of the image
to be segmented. Typically, these binary masks would obtained from
\code{\link[virustiter]{nucMask}}. Alternatively, this can be a numeric
vector for which "cutoff" values are desired.}

\item{lo}{a optional numeric value indicating the approximate minimum value.
This value will be used to search for the true saddle point. If negative,
the search will start at the lower 10 percentile of the area.}

\item{hi}{a optional numeric value for the expected upper limit of the
cutoff. If missing, the\code{quantile} will be used to determine the
upper limit.}

\item{breaks}{a number (or means) to determining the number of cells for
the optional histogram (see \code{\link[graphics]{hist}})}

\item{quant}{if \code{hi} is not provided, the area given by this
quantile will be used as the upper limit.}

\item{plot}{if \code{TRUE}, plot histogram with selected cutoff points}
}
\value{
Vector of two values recommended as cutoff values suitable for
\code{\link[virustiter]{trimMask}}
}
\description{
Find appropriate range limits for the area of nuclei defined in a
binary nuclear mask. The limits will be defined by the "saddle"
point on the low end and by the upper quantile on the high end.
}
\details{
Get the low and high cutoffs to provide to \code{\link[virustiter]{trimMask}}.
}
