% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFit.R
\name{getFit}
\alias{getFit}
\title{Fit Infectivity Results with Binomial GLM Model}
\usage{
getFit(obj, by = c("sequential", "column", "row"))
}
\arguments{
\item{obj}{A \code{data.frame} produced by \code{tally()} or by \code{score()}.}

\item{by}{A character string indicating the organization of the results
(see details.)}
}
\value{
A fitted \code{glm} model or list of models. Each fitted model has the unit
of measure attached as an attribute with the text value from \code{unit}.
}
\description{
Fit the results in a \code{data.frame} produced by either \code{tally()} or
\code{score()}.
}
\details{
This function will fit a GLM binomial model using a two-column response
where the fit is weighted by the total numbers of cases. Data are fit by
the \code{glm} function using a complementary log-log model. The data
can be organized as a single, sequential manner or by rows or columns.
If the argument \code{by} is "sequential", all the data are assumed to
represent a single assay. If \code{by} is either "row" or "column", fits
will be generated for data partitioned by row or column, respectively.

The argument \code{obj} must be produced either by \code{tally()} or by
\code{score()}. If a variable named "positive" is present, the data are
assumed to have been produced by \code{score} and the function
\code{tally} will be called before proceeding. If variables named "pos"
and "neg" are present, the data are assumed to be the product of
\code{tally}. In both cases, the data must have a variable named "moi"
or "x" representing the multiplicity of infection.
}
