% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFit.R
\name{getFit}
\alias{getFit}
\title{Fit Infectivity Results with Binomial GLM Model}
\usage{
getFit(obj, by)
}
\arguments{
\item{obj}{a \code{data.frame} produced by \code{\link{tally}}
or by \code{\link{score}}.}

\item{by}{the name of the variable in the results \code{data.frame}
that will be used to split the results before fitting
(see details.)}
}
\value{
A fitted \code{glm} model \emph{or} list of such models. Each fitted model
has the unit of measure attached as an attribute with the text value from
\code{unit}. The \code{link} function will either be "logit" or "cloglog" 
and can be identified with the \code{\link{family}} function. Note that
the fitted model uses the \code{log} of the moi values for fitting.
}
\description{
Fit the results in a \code{data.frame} produced by either \code{\link{tally}} or
\code{\link{score}}.
}
\details{
This function will fit a GLM binomial model using a two-column response
where the fit is weighted by the total numbers of cases. Data are fit by
the \code{\link{glm}} function using the \code{binomial} \code{\link{family}}
with either the default \code{"logit"} \emph{or} the complementary log-log
\code{"cloglog"} link function. The \code{"logit"} link is
appropriate for symmetrical data. The \code{"cloglog"} link is better for
asymmetrical data. The best fit will be chosen by Akaike's criteria.

If the parameter \code{by} is specified, the data frame will first be split
by this variable after coercion to a \code{factor}. In previous versions,
this option only accepted "columns" or "rows" as a potential variable. 

The argument \code{obj} must be produced either by \code{\link{tally}} or by
\code{\link{score}}. If a variable named \code{"positive"} is present, data
are assumed to have been produced by \code{\link{score}} and the function
\code{\link{tally}} will be called before proceeding. If variables named 
\code{"pos"} and \code{"neg"} are present, the data are assumed to be the
product of \code{\link{tally}}. In the case of a data frame produced by
\code{\link{score}}, the data must have a variable named \code{"moi"} or
\code{"x"} representing the multiplicity of infection. If \bold{both}
\code{"x"} and \code{"moi"} are present, an attempt will be made to choose
the more informative value and a warning will be issued. Data produced by
\code{\link{tally}} will always have a variable named \code{"x"} as the
independent variable representing the multiplicity in the \code{glm} model.
}
