% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImages.R
\name{getImages}
\alias{getImages}
\title{Get Paired Microscopic Images}
\usage{
getImages(source, type = "tiff", which.images = c(1, 2, 2),
  pattern = NULL, verbose = TRUE)
}
\arguments{
\item{source}{A character vector identifying a directory or directories
with either multilayer tiff image files \emph{or} subdirectories identified
by well with separate, paired images per well \emph{or} a character vector
of image files.}

\item{type}{A character string identifying the type of image files to parse
("tif", "tiff", "jpeg", "jpg" or "png".)}

\item{which.images}{An integer of length 2 or 3. The first two numbers indicate
the relative position of the DNA image and the target in each field. The optional
third number specifies the total number of images for each field. A value of
\code{c(1, 2)} indicates DNA first and target second. A value of \code{c(2, 1)}
indicates that the order is target first and DNA image second. A value of
\code{c(1, 2, 3)} indicates a DNA image, a target image, and a third (ignored)
image such as a phase contrast image or second fluorescent color in each set.}

\item{pattern}{Optional grep pattern as character string used by \code{list.files()}
to select image files.}

\item{verbose}{Print diagnostic messages as files are parsed and read if 
\code{TRUE}.}
}
\value{
A list of two lists containing the nuclear images (nuc) and target (tgt).
Each well or file will be represented as an element in the lists \code{nuc}
and \code{tgt}. Diagnostic messages are if \code{verbose} is \code{TRUE}.
}
\description{
Check validity and and return paired DNA and fluorescent images
appropriate for \code{parseImages()}.
}
\details{
Images specified in \code{source} will be read and checked for
the proper number of image files. This has been implemented with
\code{\link{EBImage}} and is part of a suite 
of tools to determine viral titers from fluorescent micrograph pairs. 
Typically, the first of each pair is a DNA image and the second a fluorescent 
image of the viral target.  However, this order can be changed 
with the \code{which.images} argument.

Images associated with each moi can be individual files in a
single directory where each directory is named for the well such as
\code{A1}, \code{A2}, etc. and the files within are identified as
\code{A1/file001.tif}, \code{A1/file002.tif}, etc. The well identifier
can be in upper or lower case and can contain leading zeros such as
\code{c0003/file12.tif}.

Alternatively, each group of images associated with a given moi can be
a multi-layered tiff file where the sequence of images in the file is
specified by the argument \code{which.images}.
}
\examples{
# Example with data organized by folder or well
  path.by.folder <- system.file("extdata", "by_folder", package = "virustiter")
  getImages(path.by.folder)

}
