% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTiter.R
\name{getTiter}
\alias{getTiter}
\alias{getEC63}
\title{Get Titer from Fitted Model}
\usage{
getTiter(fm, signif = 3)

getEC63(fm)
}
\arguments{
\item{fm}{Fitted model or list of fitted models generated by
\code{\link{getFit}}.}

\item{signif}{\code{Integer} number of significant digits to include
with \code{getTiter} (default of 3)}
}
\value{
For \code{getTiter}, either a named numeric vector or matrix with the
estimated titer in IU per ml with +/- 95\% confidence intervals.

For \code{getEC63}, either a named numeric vector or matrix with the estimated
volume corresponding to 1 infectious unit with +/- 95\% confidence intervals.
}
\description{
Extract the viral titer \emph{or} the volume required for one
infectious unit with confidence intervals from a fit object or
a list of fit objects generated by \code{\link{getFit}}.
}
