% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTiter.R
\name{getTiter}
\alias{getTiter}
\title{Get Titer from Fitted Model}
\usage{
getTiter(fm)
}
\arguments{
\item{fm}{Fitted model or list of fitted models generated by
\code{\link{getFit}}.}
}
\value{
Either a named numeric vector or matrix with the estimate and 95\% low
and hi confidence intervals.
}
\description{
Extract the viral titer with confidence intervals from a fit or
list of fits generated by \code{getFit}.
}
