% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTiter.R
\name{getTiter}
\alias{getTiter}
\alias{getEC63}
\title{Get Titer from Fitted Model}
\usage{
getTiter(fm, digits = 3, level = 0.95)

getEC63(fm, level)
}
\arguments{
\item{fm}{fitted model \emph{or} list of fitted models generated by
\code{\link{getFit}}.}

\item{digits}{\code{Integer} number of significant digits to report
(default = 3)}

\item{level}{this specifies the confidence level (default of 0.95).
Use \code{NULL} to return only the titer (or EC63 value).}
}
\value{
For \code{getTiter}, a named numeric vector with the estimated titer in
IU per ml.

For \code{getEC63}, a named numeric vector with the volume corresponding to 1
1 infectious units.

If \code{level} was provided, the return value is a matrix with additional
columns providing the desired +/- confidence intervals.

For \code{getEC63}, either a named numeric vector or matrix with the estimated
volume corresponding to 1 infectious unit with +/- 95\% confidence intervals.
}
\description{
Extract the viral titer \emph{or} the volume required for one
infectious unit with confidence intervals from a fit object or
a list of fit objects generated by \code{\link{getFit}}.
}
