% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTiter.R
\name{getTiter}
\alias{getTiter}
\alias{getEC63}
\title{Get Titer from Fitted Model}
\usage{
getTiter(fm, level = 0.95, digits = 3)

getEC63(fm, level = 0.95)
}
\arguments{
\item{fm}{fitted model \emph{or} list of fitted models generated by
\code{\link{getFit}}.}

\item{level}{this specifies the confidence level (default of 0.95).
Use \code{NULL} to return only the titer (or EC63 value).}

\item{digits}{\code{Integer} number of significant digits to report
(default = 3)}
}
\value{
For \code{getTiter}, a named numeric vector with the estimated titer in
IU per ml.

For \code{getEC63}, a named numeric vector with the volume corresponding to
one infectious unit.

If \code{level} was provided, the return value is a matrix with additional
columns providing the desired +/- confidence intervals.
}
\description{
Extract the viral titer \emph{or} the volume required for one
infectious unit with confidence intervals from a fit object or
a list of fit objects generated by \code{\link{getFit}}.
}
