% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVal.R
\name{getVal}
\alias{getVal}
\title{Get Value from ComputeFeatures.xxx Function}
\usage{
getVal(mask, target, val = "b.mean", FUN = computeFeatures.basic,
  simplify = TRUE)
}
\arguments{
\item{mask}{An object mask or list of objects masks with
connected pixels having the same integer value.}

\item{target}{A fluorescent \code{Image} object or list of
\code{Image}s corresponding to the objects in \code{mask}.}

\item{val}{A character string identifying the parameter
to return from the computeFeatures function \code{FUN}. The
default value of \code{"b.mean"} returns the mean intensity.}

\item{FUN}{A \code{\link[EBImage]{computeFeatures}} function to be
applied over \code{mask} and \code{target}, default of 
\code{computeFeatures.basic}.}

\item{simplify}{If \code{TRUE} (default), the result will be
collapsed into a vector. Otherwise, a vector for each member of 
the list \code{mask} and \code{target} will be returned.}
}
\value{
A vector if \code{simplify} is\code{TRUE} or a list of values obtained
from \code{FUN}.
}
\description{
Extract a value from one of the 'computeFeatures' family of
functions using a mask or a list of masks and target image
or list of images.
}
\details{
Objects identified in \code{mask} will be projected onto \code{target} 
and quantified with the function specified in \code{FUN}. The single 
value specified by \code{val} will be returned as a list if \code{
simplify == FALSE} or as a single vector if \code{simplify == TRUE}.

Possible functions include \code{computeFeatures.basic, 
computeFeatures.shape, computeFeatures.moment}, and \code{
computeFeatures.haralick}. See the help page for
\code{\link[EBImage]{computeFeatures}} for possible values for
\code{FUN} and \code{val}.
}
