% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVal.R
\name{getVal}
\alias{getVal}
\title{Get Value from ComputeFeatures.xxx Function}
\usage{
getVal(mask, ref = NULL, val = "b.mean", FUN = NULL, simplify = TRUE, ...)
}
\arguments{
\item{mask}{An object mask or list of objects masks with
connected pixels having the same integer value.}

\item{ref}{A fluorescent \code{Image} object or list of
\code{Image} objects corresponding to the objects in \code{mask}
or \code{NULL} if a reference image is not required. If \code{ref}
is a character string, it is \emph{assumed} to be the value
for \code{val} and will be used in place of the default.}

\item{val}{A character string identifying the parameter
to return from the \code{computeFeatures} function assigned
to \code{FUN}. The default value of \code{"b.mean"} returns the
mean intensity with the
\code{\link[EBImage:computeFeatures]{computeFeatures.basic}}
function.}

\item{FUN}{A \code{\link[EBImage]{computeFeatures}} function to be
applied over \code{mask} and \code{ref}. The default of 
\code{NULL} uses the character string in \code{val} to select
the appropriate function. If this function is specified,
the value in the argument \code{val} must be an exact match
to one of the values returned by the specified function.}

\item{simplify}{If \code{TRUE} (default), the result will be
collapsed into a vector. Otherwise, a vector for each member of 
the list \code{mask} and \code{ref} will be returned.}

\item{...}{additional parameters are passed to \code{FUN}.}
}
\value{
A vector if \code{simplify} is \code{TRUE} or a list of values obtained
from \code{FUN}.
}
\description{
Extract a value through one of the \code{computeFeatures} family of
functions using an object mask and reference image or a list of each.
}
\details{
Objects identified in \code{mask} will be projected onto \code{ref} 
and quantified with the function specified in \code{FUN}. The single 
value specified by \code{val} will be returned as a list if \code{
simplify == FALSE} or as a single vector if \code{simplify == TRUE}.

Common usages include getting the mean intensity from an object mask
and reference image or getting the area of objects from an object mask
as shown below. The code has been accelerated by a factor of 8 
for getting the area by using the \code{\link{tabulate}} function
rather than the \code{computeFeatures.shape} function.

\preformatted{
# To extract the mean intensity:
  mfi <- getVal(mask, ref) # uses default parameter "b.mean"

# To extract object area, collapsed to a single vector:
  area <- getVal(mask, val = "area") # no need 'ref'

# A complex example of Haralick difference entropy at
# at a Haralick scale of 4 pixels
  den <- getVal(mask, ref, val = "h.den.s4",
    FUN = computeFeatures.haralick, haralick.scales = 4)
}

The character string in \code{val} is used as a \code{grep} pattern to
search among the parameters returned by the \code{computeFeatures}
functions. Consequently, a partial match such as "ecc" is sufficient
for a parameter such as "m.eccentricity". However, because "mean" occurs
in "b.mean" and "s.radius.mean", a partial match to "mean" does not
identify a unique parameter. The unique parameter is used to assign the
value to \code{FUN}. This value can be explicitly assigned as one of
\code{\link[EBImage:computeFeatures]{computeFeatures.basic}}, 
\code{\link[EBImage:computeFeatures]{computeFeatures.shape}}, 
\code{\link[EBImage:computeFeatures]{computeFeatures.moment}} and 
\code{\link[EBImage:computeFeatures]{computeFeatures.haralick}}.

See the \code{EBImage} help page for appropriate values for
\code{FUN} and \code{val} but \strong{note} that "\code{basic quantiles}" 
parameters are mislabeled with an extra '\code{0}'. For example, the 5th
percentile and 95th percentile values are named "\code{b.q005}"
and "\code{b.q095}", respectively.
}
