% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePdata.R
\name{mergePdata}
\alias{mergePdata}
\title{Merge Phenotype and Image Data}
\usage{
mergePdata(phenoData, imageData, moi = c("moi", "x"))
}
\arguments{
\item{phenoData}{A \code{data.frame} containing variables for the multiplicity,
typically \code{moi} and either \code{well} or \code{file} associated with
the unique multiplicity.}

\item{imageData}{A \code{data.frame} generated by \code{parseImages()}.}

\item{moi}{Character name for multiplicity variable, either "moi" or "x".}
}
\value{
Merged \code{data.frame} with harmonized well information (if appropriate)
and additional data provided in \code{phenoData}.
}
\description{
Merge "phenotype" data describing the samples in an image data.frame.
}
\details{
This function performs a \code{data.frame} merge with additional checks.
}
