% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePdata.R
\name{mergePdata}
\alias{mergePdata}
\title{Merge Phenotype and Image Data}
\usage{
mergePdata(
  phenoData,
  imageData,
  moi = c("moi", "x"),
  stringsAsFactors = TRUE,
  formatString = NULL
)
}
\arguments{
\item{phenoData}{A \code{data.frame} containing variables for the multiplicity,
typically \code{moi} and either \code{well} or \code{file} associated with
each unique multiplicity.}

\item{imageData}{A \code{data.frame} generated by \code{\link{parseImages}}.}

\item{moi}{Character name for multiplicity variable, either "moi" or "x".}

\item{stringsAsFactors}{Convert string variables to factors if \code{TRUE}
(default value). Note that if present, \code{column} and \code{row} are
preserved as factors regardless of this conversion.}

\item{formatString}{Character string as \code{\link{sprintf}} format for the
column; default value of \code{NULL} pads the column value with zeros. 
See details for fixing a mismatch between imageData and phenoData}
}
\value{
Merged \code{data.frame} with harmonized well information (if appropriate)
and additional data provided in \code{phenoData}. The variables \code{row},
\code{column}, and \code{well} are expected to be factors.
}
\description{
Merge "phenotype" data describing the samples in an image data.frame.
}
\details{
The format for well names (and column names) may differ between the imageData
and phenoData. An "Error in mergePData(pd, df)" might report "A1, A2, <...>
are missing from pd." This can happen if, for example, the imageData uses only
a few of the possible 96 wells in phenoData. Change the \code{formatString} 
parameter to a value such as "\%02d" to force a consistent format. This value
is passed to the \code{\link{well.info}} as the parameter \code{format}.

This function performs a \code{data.frame} merge with additional checks. It
does not (yet) respect the 'plate' variable and only merges according to the
\code{well} or \code{file} variables. For now, it is necessary to split
the phenoData and imageData by \code{plate}, merge each separately,
and then join the data back together. The following might work with phenoData
in \code{pd} and imageData in \code{df}:

\preformatted{
  g <- pd$plate
  pd <- split(pd, g)
  df <- split(df, df$plate)
  df <- Map(mergePdata, pd, df)
  df <- do.call(rbind, df)
  rownames(df) <- NULL
  pd <- unsplit(pd, g)
}
}
