% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucMask.R
\name{nucMask}
\alias{nucMask}
\title{Generate Nuclear Mask from DNA Image}
\usage{
nucMask(dna, width = 36, offset = 0.05, size = 2, sigma = 2,
  radius = NULL, gamma = 1, border = 0)
}
\arguments{
\item{dna}{Fluorescent DNA \code{Image} object \emph{or} a character
vector representing a path or paths to DNA image file(s).}

\item{width}{Largest nuclear width (diameter) used as width parameter
for \code{thresh2}.}

\item{offset}{Offset parameter for \code{thresh2}. Use 0.05
for typical images, use 0.01 for low contrast images.}

\item{size}{Radius for \code{medianFilter}, integer. Use 2 for typical
images, use 0 to skip \code{medianFilter}.}

\item{sigma}{Standard deviation for \code{gblur}, use 2 for typical
images, use 5 for finely detailed images.}

\item{radius}{Radius for \code{gblur}, use default of 2 * ceiling(3 * sigma)
+ 1 for typical images, use numbers smaller than the default of 13 for
images with smaller nuclei.}

\item{gamma}{Exponent for \code{DNA^gamma} transformation.}

\item{border}{Exclude objects within this many pixels from the edge.}
}
\value{
A single object holding an integer \code{Image} mask for each DNA image.
A mask for a single image of dimension \code{dim(dna) = c(nx, ny)} will be
returned with a third dimension equal to 1. \code{dim(mask) = c(x, y, 1)}.
}
\description{
Generate and return an integer \code{Image} mask from a fluorescent DNA image.
}
\details{
Generate an integer object mask representing segmented nuclei. The
argument can be either a monochrome nuclear fluorescent image of one
or more dimensions \emph{or} a character vector of path(s) to such images.

Optimal conditions may need to be found by empirically adjusting the
arguments, especially \code{width} and \code{offset}.

The image or images will be processed sequentially by (1) an optional gamma 
transformation, (2) normalization, (3) \code{medianFilter()} with argument 
\code{size} if \code{size} is non-zero, (4) \code{gblur()} with arguments 
\code{sigma} and \code{radius}, (5) \code{thresh2()} with arguments 
\code{width} and \code{offset}, (6) \code{fillHull()}, (7) \code{distmap()}, 
(8) \code{watershed()} and then (9) optionally removing objects along
the edge of the image.
}
\examples{
  f.example <- system.file("extdata", "by_folder/b2/file003.tif",package = "virustiter")
  xm0 <- nucMask(f.example)
  max(xm0) # total number of nuclei
  xm4 <- nucMask(f.example, border = 4)
  max(xm4) # total number of nuclei
  opar <- par(mfrow = c(1, 2))
  plot(colorLabels(xm0))
  plot(colorLabels(xm4))
  par(opar)

}
