% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucMask.R
\name{nucMask}
\alias{nucMask}
\title{Generate Nuclear Mask from DNA Image}
\usage{
nucMask(dna, width = 36, offset = 0.05, size = 2, sigma = 2,
  radius = NULL, gamma = 1)
}
\arguments{
\item{dna}{Fluorescent DNA \code{Image} or list of fluorescent DNA
\code{Image}s.}

\item{width}{Maximum nuclear diameter (in pixels) to be used as
the \code{width} parameter for \code{thresh2}.}

\item{offset}{Offset parameter for \code{thresh2}. Use 0.05
for typical images, use 0.01 for low contrast images.}

\item{size}{Radius (in pixels) for \code{medianFilter} as an integer.
Use 2 for typical images, use 0 to skip \code{medianFilter}.}

\item{sigma}{Standard deviation for \code{gblur}, use 2 for typical
images, use 5 for finely detailed images.}

\item{radius}{Radius for \code{gblur}, use default value of
\code{2 * ceiling(3 * sigma) + 1} for typical images, use numbers
smaller than the default of 13 for images with smaller nuclei.}

\item{gamma}{Exponent used for \code{DNA^gamma} transformation.}
}
\value{
An integer \code{Image} mask or list of integer \code{Image} masks.
Note that masks single images of dimension \code{dim(dna) = c(nx, ny)}
will be returned with a third dimension set to 1. (\code{dim(mask) =
c(nx, ny, 1)}.
}
\description{
Generate an integer \code{Image} mask from a fluorescent DNA image.
}
\details{
Generate an integer object mask (or list of masks) representing
segmented nuclei. The argument \code{nuc} must be a grayscale nuclear image
of one or more dimensions \emph{or} a list of such images.

Optimal conditions may need to be found by empirically adjusting the
arguments, especially \code{width} and \code{offset}.

The image or images will be processed sequentially by (1) an optional gamma 
transformation, (2) normalization, (3) \code{medianFilter()} with argument 
\code{size} if \code{size} is non-zero, (4) \code{gblur()} with arguments 
\code{sigma} and \code{radius}, (5) \code{thresh2()} with arguments 
\code{width} and \code{offset}, (6) \code{fillHull()}, (7) \code{distmap()}, 
and (8) \code{watershed()}.
}
\examples{
  f.ex <- system.file("extdata", "by_folder/b2/file003.tif",package = "virustiter")
  nuc.ex <- readImage(f.ex) # single nuclear image
  xm0 <- nucMask(nuc.ex)
  plot(colorLabels(xm0))

}
