% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2p.R
\name{p2p}
\alias{p2p}
\title{Interactively Measure Point-to-Point Distances}
\usage{
p2p(n = 512, col = "magenta", type = "o", pch = 3, ...)
}
\arguments{
\item{n}{Maximum number of pairs to measure, integer.}

\item{col, type, pch}{Values handed to \code{\link{locator}}.}

\item{...}{Additional values handed to \code{\link{locator}}.}
}
\value{
Distances between points with locator in pixels.
}
\description{
Use the \code{\link{locator}} function in base graphics to interactively
measure distances between pairs of points on an existing image.
End the interaction pressing the \strong{Esc} key \emph{or}
control-clicking.
}
