% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseImages.R
\name{parseImages}
\alias{parseImages}
\title{Extract Fluorescent Intensities from Paired Microscopic Images}
\usage{
parseImages(path, type = "tiff", which.images = c(1, 2, 2),
  pattern = NULL, args.nucMask = NULL, args.trimMask = NULL,
  cellMask.flag = FALSE, equalize = FALSE)
}
\arguments{
\item{path}{A character vector of directory with either multilayer tiff
image files \emph{or} subdirectories identified by well with separate,
paired images per well.}

\item{type}{A character string identifying the type of image files to parse
("tif", "tiff", "jpeg", "jpg" or "png".)}

\item{which.images}{An integer of length 2 or 3. The first two numbers indicate
the relative position of the DNA image and the targetin each field. The optional
third number specifies the total number of images for each field. A value of
c(1, 2) indicates DNA first and target second. A value of c(2, 1) indicates that
the order is target first and DNA image second. A value of c(1, 2, 3)
indicates a DNA image, a target image, and a third (ignored) image such as a
phase contrast image or second fluorescent color in each set.}

\item{pattern}{Optional grep pattern as character string used by \code{list.files()}
to select image files.}

\item{args.nucMask}{A list of arguments passed to \code{nucMask()}.}

\item{args.trimMask}{A list of arguments passed to \code{trimMask()} such as
\code{cutoff} or \code{k}. If this value is \code{NA}, no trimming is performed.}

\item{cellMask.flag}{If this \code{logical} value is \code{TRUE}, the default
nuclear mask will be used to generate a mask with \code{cellMask()}. This
"cellular" mask will be used to measure fluorescence in the target image.}

\item{equalize}{\emph{If the fluorescent target images have more background
pixels than foreground pixels} and if the background varies significantly
from image to image, this {logical} parameter can be set to
\code{TRUE} in order to equalized the fluorescent images by subtracting the
median value after applying a median filter and gaussian blur using the
function \code{bnormalize()}.}
}
\value{
A data.frame of processed image data. \strong{All} data.frames will have the
following variables:
\describe{
  \item{\code{directory}}{Path to enclosing folder.}
  \item{\code{frame}}{Image sequence (1, 2, 3, ...)}
  \item{\code{xm, ym}}{Center of mass (in pixels) for nucleus.}
  \item{\code{area}}{Area of mask (nuclear or cell).}
  \item{\code{dna}}{Mean fluorescence intensity for DNA stain,
    typically not meaningful with over-exposed images.}
  \item{\code{mfi}}{Mean fluorescence intensity for signal of interest,
    measured with selected mask.}
}
Results from data organized by \strong{well} will also include:
\describe{
  \item{\code{well}}{Harmonized well identifier. See \code{well.info()} function.}
  \item{\code{row}}{Row identifier ("A", "B", "C", etc.) as a factor.}
  \item{\code{column}}{Column number as a factor.}
}
while results from data organized as \strong{stacks} (multi-layered
tiff files) will include:
\describe{
  \item{\code{file}}{The file name as a factor.}
}
}
\description{
Extract the mean fluorescence intensity of DNA and a second fluorescent
target image for individual cells from paired images.
}
\details{
This is the core function that reads and parses image data in a suite of tools
implemented with \code{\link{EBImage}} which has been developed to determine
viral titers from sets of fluorescent micrographs. Typically, the images are
acquired as paired images where the first of each pair
is a DNA image and the second a fluorescent image of the viral target.
Because individual cells are identified by the nuclear stain, it
\emph{may} be beneficial to collect overexposed DNA images.

This function was developed to process fluorescent virus titers
performed in multi-well plates and is designed to parse images
collected at different multiplicities of infection or moi. The
moi is expressed as virions (VP) \emph{or} infectious
units (IU) \emph{or} volume (ml, ul, nl) per cell and is added to the
data results of this function with the \code{mergePdata} function.
The nuclear (typically DAPI) image file is expected to precede the
corresponding viral antigen image file but this order can be changed with
the \code{which.images} argument.

Images associated with each moi can be individual files in a
single directory where each directory is named for the well such as
\code{A1}, \code{A2}, etc. and the files within are identified as
\code{A1/file001.tif}, \code{A1/file002.tif}, etc. The well identifier
can be in upper or lower case and can contain leading zeros such as
\code{c0003/file12.tif}.

Alternatively, each group of images associated with a given moi can be
a multi-layered tiff file where the sequence of images in the file is
specified by the argument \code{which.images}.

If the fluorescent images have \emph{more} background pixels
than foreground
pixels and if the images have background values that vary from image
to image or have significant noise, the argument \code{equalize} can be set
to \code{TRUE} to smooth the images by sequentially
\emph{modifying the values in each} image with a
median filter of radius 2, a gaussian blur of radius 2 followed by
subtracting the median value for each image and adding an offset of 0.05.
\emph{This cannot be used for a data set that has images with a large fraction
of positive cells, such as the example data set.}
}
\examples{
# Note that execution of these examples can be rather slow...
  path.by.folder <- system.file("extdata", "by_folder", package = "virustiter")
  df.by.folder <- parseImages(path.by.folder)
  head(df.by.folder)[-1] # drop potentially log directory name

  path.by.stack <- system.file("extdata", "by_stack", package = "virustiter")
  df.by.stack <- parseImages(path.by.stack)
  head(df.by.stack)[-1] # drop potentially log directory name

# plots
  opar <- par(mfrow = c(1,2))
  plot(log(mfi) ~ area, df.by.folder, main = "By Folder", las = 1)
  plot(log(mfi) ~ area, df.by.stack, main = "By Stack", las = 1)
  par(opar)

}
