% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDens.R
\name{plotDens}
\alias{plotDens}
\title{Show Density Plot of Mean Fluorescence Intensity with Background Cutoff}
\usage{
plotDens(df, bgnd, param = "mfi", panel, adjust = 1, log = TRUE,
  by = NULL, mult = NULL, main = NULL, as.table = TRUE,
  layout = NULL, ...)
}
\arguments{
\item{df}{Annotated \code{data.frame} with imaging results.}

\item{bgnd}{Numeric vector of background values. If missing, \code{\link{getBgnd}}
will be called with parameters \code{by, param, mult,} and \code{log}.}

\item{param}{Character string identifying the variable to be analyzed. Also
passed to \code{\link{getBgnd}} if required.}

\item{panel}{Optional character string defining the \code{lattice} panels,
typically \code{"well"} or \code{"file"}.}

\item{adjust}{Numeric value controlling bandwith, passed to the
\code{\link{density}} function.}

\item{log}{Optional \code{logical} or \code{numeric} value to transform
\code{'param'} values. Also passed to \code{\link{getBgnd}} if required.}

\item{by, mult}{Additional parameters passed to \code{\link{getBgnd}} if required.}

\item{main}{Optional character string to serve as plot title. If \code{NULL},
the system date will be used.}

\item{as.table}{A \code{logical} value passed to \code{densityplot}.}

\item{layout}{An optional numeric vector to specify layout of densityplot,
passed to \code{densityplot}.}

\item{...}{Additional arguments passed to \code{densityplot}.}
}
\value{
The plot is returned as an invisible \code{lattice} object.
}
\description{
Display a \code{densityplot} of each well or file with \code{lattice}
graphics showing the selected background cutoff values.
}
\details{
A densityplot of image intensity in \code{'param'} is plotted
with the selected background cutoff similar to the function
\code{\link{plotHist}}. Both functions can be used to examine the uniformity
of results from an imaging experiment and to interactively check the
paramaters passed to \code{\link{getBgnd}} to determine a suitable background
cutoff value. The background values are incorporated into the strip labels.
}
