% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDens.R
\name{plotDens}
\alias{plotDens}
\title{Show Density Plot of Mean Fluorescence Intensity with Background Cutoff}
\usage{
plotDens(df, bgnd, by = c("default", "well", "file", "row", "column"),
  smooth = 1, mult = 2.5, log = TRUE, main = NULL, as.table = TRUE,
  param = "mfi", return.plot = FALSE, ...)
}
\arguments{
\item{df}{Annotated \code{data.frame} with imaging results.}

\item{bgnd}{Numeric vector of bgnd values. If missing, \code{getBgnd()}
will be called with default parameters.}

\item{by}{Character vector indicating grouping where "default" will use
\code{well} if present or \code{file}. This value is also passed to
\code{getBgnd()} if necessary.}

\item{smooth}{Numeric value passed to the \code{density()} function.}

\item{mult}{Numeric value passed to \code{getBgnd()} to scale bgnd.}

\item{log}{A \code{logical} value passed to \code{getBgnd()}.}

\item{main}{Optional character string to serve as plot title.}

\item{as.table}{A \code{logical} value passed to \code{histogram()}.}

\item{param}{Name of the variable to be analyzed as a character string; 
typically "mfi" or "val".}

\item{...}{Additional arguments handed to \code{histogram()}.}
}
\value{
The plot is returned as an invisible \code{lattice} object.
}
\description{
Display a \code{densityplot} of each well or file with \code{lattice} graphics
showing the selected background cutoff value or values.
}
\details{
This presents similar representation of the data as \code{plotHist} and can
be used to examine the uniformity of results from an imaging experiment and
to iteratively check the \code{mult} argument provided to \code{getBgnd()}.
The background values are incorporated into the strip labels.
}
