% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFit.R
\name{plotFit}
\alias{plotFit}
\title{Plot Titer Fit from GLM Model}
\usage{
plotFit(fm, by = c("none", "column", "row"), line.col = 2, ...)
}
\arguments{
\item{fm}{Fitted model or list of fitted models from \code{getFit()}.}

\item{by}{Character vector indicating that fits are organized as a
single fit ("none") or by "column" or "row".}

\item{line.col}{Color for best-fit line (appropriate for \code{par("col")}.}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\value{
No value is returned. This function is called for the side-effect of 
producing a plot (or plots).
}
\description{
Prepare a diagnositc plot by column, row, or none.
}
\details{
Base graphics are used to prepare the plot where the first argument is
either a single fitted model or a list of fitted models from
\code{getFit()}. The function calls \code{getTiter()} to obtain the fit
and confidence intervals.
}
