% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFit.R
\name{plotFit}
\alias{plotFit}
\title{Plot Titer Fit from GLM Model}
\usage{
plotFit(fm, line.col = 2, main = NULL)
}
\arguments{
\item{fm}{Fitted model or list of fitted models from \code{\link{getFit}}.}

\item{line.col}{Color for best-fit line appropriate for \code{par("col")}.}

\item{main}{List or character vectors appropriate for the plot title
of the same length as \code{fm}. If \code{NULL}, the current date will
be used.}
}
\value{
No value is returned. This function is called for the side-effect of 
producing a plot (or plots).
}
\description{
Prepare a diagnostic plot of viral titer GLM fit.
}
\details{
Base graphics are used to prepare the plot where the first argument is
either a single fitted model or a list of fitted models from
\code{\link{getFit}}. The function calls \code{\link{getTiter}} to obtain the fit
and confidence intervals.
}
