% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHist.R
\name{plotHist}
\alias{plotHist}
\title{Show Cutoff with Histogram of Mean Fluorescence Intensity}
\usage{
plotHist(df, cut, by = c("default", "well", "file", "row", "column"),
  mult = 5, log = TRUE, param = "mfi", main = NULL, as.table = TRUE,
  layout = NULL, ...)
}
\arguments{
\item{df}{Annotated \code{data.frame} with imaging results.}

\item{cut}{Numeric vector of cutoff values. If missing, \code{getCut} will be
called with default parameters.}

\item{by}{Character vector indicating grouping where "default" will use
\code{well} if present or \code{file}. This value is also passed to
\code{getCut} if necessary.}

\item{mult}{Numeric value passed to \code{getCut} to scale cutoff.}

\item{log}{A \code{logical} value passed to \code{getCut}.}

\item{param}{Name of the variable to be analyzed as a character string; 
typically "mfi" or "val".}

\item{main}{Optional character string to serve as plot title.}

\item{as.table}{A \code{logical} value passed to \code{histogram}.}

\item{layout}{An optional numeric vector to specify layout of histogram,
passed to \code{histogram}.}

\item{...}{Additional arguments handed to \code{histogram}.}
}
\value{
The plot is returned as an invisible \code{lattice} object.
}
\description{
Display a histogram of each well or file with \code{lattice} graphics showing
the selected background cutoff value or values.
}
\details{
This presents similar representation of the data as \code{plotDensity} and
can be used to examine the uniformity of results from an imaging experiment
and to iteratively check the \code{mult} argument provided to \code{getCut}.
}
