% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPlate.R
\name{plotPlate}
\alias{plotPlate}
\title{Plot Results of an Entire Plate by Well or File}
\usage{
plotPlate(df, size = NULL, cex = 1/2, alpha = 1/2, main = NULL,
  invert.y = TRUE, ...)
}
\arguments{
\item{df}{Annotated \code{data.frame} with imaging results.}

\item{size}{Number of wells for plate, if \code{NULL}, the size will be guessed."}

\item{cex}{Numeric value as factor by which to expand the plot symbol.}

\item{alpha}{Transparency factor applied to plotting symbols.}

\item{main}{Optional character string to serve as plot title.}

\item{invert.y}{A \code{logical} value to invert y coordinates.}

\item{...}{Additional arguments handed to \code{xyplot} including subset.}
}
\value{
The plot is returned as an invisible \code{lattice} object.
}
\description{
Display an entire plate with lattice graphics grouped by well or file.
}
