% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh2.R
\name{thresh2}
\alias{thresh2}
\title{Adaptive Threshold with a Moving Disc Window}
\usage{
thresh2(x, width, offset, boundary = c("replicate", "circular"))
}
\arguments{
\item{x}{\code{Image} object or array.}

\item{width}{Half-width of the moving disc window.}

\item{offset}{Thresholding value from the averaged value.}

\item{boundary}{Behavior at image border, where default is to extend the 
image beyond the borders, which differs from the default for
\code{filter2}.}
}
\description{
Threshold an image using a disc rather than rectangle.
}
\details{
This function returns the binary image generated by a local thresholding 
algorithm using a moving disc.
}
\examples{

  x <- readImage(system.file("extdata", "by_stack/file001.tif", package = "virustiter"))
  xb <- normalize(gblur(x, 2))
  xt <- thresh2(xb, width = 12)
  plot(combine(xb, xt), all = TRUE)

}
