% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimMask.R
\name{trimMask}
\alias{trimMask}
\title{Remove small and large objects from an object mask}
\usage{
trimMask(mask, cutoff = NULL, k = c(1.5, 3), reenumerate = TRUE)
}
\arguments{
\item{mask}{object mask with connected pixels having the same
integer value}

\item{cutoff}{optional integer value of length 2 specifying the lower 
and upper absolute cutoff values}

\item{k}{numeric value of length 2 specifying the lower and upper multiplier 
to determine the cutoff from the \code{mean} and \code{mad} of the area}

\item{reenumerate}{reenumerate the objects in the trimmed mask}
}
\value{
Object mask with small and large objects removed.
}
\description{
Remove objects with area outside of lower and upper cutoff values
or by a lower and upper multiplier applied to the \code{mad} of the area.
}
\details{
If \code{cutoff} is specified, objects smaller than \code{cutoff[1]} and 
larger than \code{cutoff[2]} will be removed otherwise, objects smaller 
than \code{mean(area) - k[1]*mad(area)} and larger than \code{mean(area)+ 
k[2]*mad(area)} will be removed. The mask will be reenumerated if that 
parameter is \code{TRUE}.
}
\examples{

x <- readImage(system.file("extdata", "by_stack/file001.tif",
package = "virustiter"))
xb <- normalize(gblur(x, 2))
xt <- thresh(xb)
xm <- bwlabel(xt)
xm2 <- trimMask(xm)
xm3 <- trimMask(xm, cutoff = c(1, 20))
xm4 <- trimMask(xm, cutoff = c(100, 250))
plot(combine(xm, xm2, xm3, xm4), all = TRUE)

}
