% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimMask.R
\name{trimMask}
\alias{trimMask}
\title{Remove Objects by Size, Proximity to Edge and Eccentricity}
\usage{
trimMask(
  mask,
  cutoff = FALSE,
  k = c(1.5, 3),
  border = 0,
  brush = 0,
  ecc.max = 1,
  reenumerate = TRUE
)
}
\arguments{
\item{mask}{Object mask or list of masks with connected pixels having
the same integer value.}

\item{cutoff}{Optional integer value of length 1 or 2 specifying the
limits for the area in pixels. If only one value is provided, it
is assumed to be the lower limit. A value of \code{FALSE} or
\code{NA} prevents size exclusion from occurring. A value of
\code{NULL} makes use of the multiplier parameter \code{k} to determine 
the cutoff limits. Either of the two values in \code{cutoff} can be
specified as \code{NA} to use the multiplier parameter for that position.}

\item{k}{Numeric value of length 2 specifying the lower and upper
multiplier to determine the cutoff from the \code{\link{median}} and
\code{\link{mad}} of the area if \code{cutoff} is \code{NULL}.}

\item{border}{Objects within this many pixels of the edge will be excluded.}

\item{brush}{After coercion to the nearest odd integer, values > 0
\code{\link[EBImage:morphology]{dilate}} the mask whereas values < 0
\code{\link[EBImage:morphology]{erode}} the mask.}

\item{ecc.max}{Exclude objects with elliptical eccentricity greater than
this value.}

\item{reenumerate}{Logical value (default of TRUE) passed to rmObjects()
to re-enumerate the objects after trimming.}
}
\value{
Object mask or list of masks with objects removed and re-enumerated.
}
\description{
Remove large and small objects from an integer \code{Image} mask,
remove objects that are near the edge of the image, and remove
eccentric objects.
}
\details{
For each non-\code{NA} value in \code{cutoff}, objects smaller than
\code{cutoff[1]} and larger than \code{cutoff[2]} will be removed. 
Otherwise, objects smaller than \code{median(area) - k[1]*mad(area)} and
larger than \code{median(area)+ k[2]*mad(area)} will be removed. Objects
that are within \code{border} pixels of the edge will be removed. Objects
that have eccentricity greater than \code{ecc.max} will be removed. A circle
has eccentricity of 0 and a straight line has eccentricity of 1. Note 
that EBImage provides a poor approximation of this value. The final mask
will be dilated or eroded if \code{brush} is non-zero using a disc-shaped
brush with size determined by the adjusted value of \code{brush}.
}
\examples{

  x <- readImage(system.file("extdata", "by_folder/a1/file001.tif", package = "virustiter"))
  xm <- nucMask(x)
  xm2 <- trimMask(xm, cutoff = c(200, 400))
  xm3 <- trimMask(xm, border = 24)
  xm4 <- trimMask(xm, ecc.max = 0.75)
  plot(colorLabels(combine(xm, xm2, xm3, xm4)), all = TRUE)
  sapply(list(xm, xm2, xm3, xm4), max) # how many remain?

}
