% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well.info.R
\name{well.info}
\alias{well.info}
\title{Extract Prefix, Row and Column from Well Label}
\usage{
well.info(w, format = "\%02d", upper = TRUE)
}
\arguments{
\item{w}{Name of the well (coerced as a character).}

\item{format}{Character string as \code{sprintf} format for the column portion.}

\item{upper}{\code{logical} value to use upper case when \code{TRUE}.}
}
\value{
A named list of length four:
\itemize{
	\item prefix, the optional prefix
	\item well, uniform character representation of the well
	\item row, single character for the row
	\item column, column number as an integer
}
}
\description{
Create a uniform representation for a well with an optional 
prefix.
}
\details{
Parse the value in \code{w} into an optional arbitrary prefix followed by a 
single character for the well, followed by the column number formatted
as per \code{format}. The code will fail if the row value is not one of the 26
letters or if the column number is not in the range of 1 to 384.
}
\examples{
  well.info("2a6") # default settings
  well.info("2a6", format = "\%d", upper = TRUE)
  well.info("b92", format = "\%05d")

}
